efi/fdt: Add SDM850 UEFI workaround
The UEFI implementation on SDM850 has a number of bugs which prevent the
efi stub from properly booting the Linux kernel, and attempting to do so
results in system crashes.  Work around these issues by manipulating the
MMU in a way to avoid the faulty scenarios during the final steps of
shutting down UEFI so that the system can successfully get into the kernel.

Signed-off-by: AArch64 Laptops <aarch64laptops@gmail.com>