scsi: ufs: ufshcd-pltfrm: Reduce burden when !OF
When OF (Device Tree) is enabled, ufshcd_parse_regulator_info() calls
into ufshcd_populate_vreg() 4 times, only to be rejected each time.
Seeing as ufshcd_parse_regulator_info() is the only caller of
ufshcd_populate_vreg() it's better to move the check for !OF into
there instead.  This saves duplicating lines in the kernel log and
saves unnecessary cycles.

Signed-off-by: Lee Jones <lee.jones@linaro.org>