soc: qcom: add acpi-urs glue driver support
It adds an ACPI URS glue driver to populate platform device for the
child node of URS in ACPI DSDT table, e.g. the USB0 on SDM850.  As the
memory resource resides in URS node and interrupt lies in USB0 node, it
needs to create resources for the platform device by composing these two
pieces.

Signed-off-by: Shawn Guo <shawn.guo@linaro.org>